Fix compilation on Windows

+++ src/quickjs_utils/typedarrays.rs
@@ -78,7 +78,7 @@ pub unsafe fn new_array_buffer(
     #[cfg(target_pointer_width = "64")]
     let length = buf.len();
     #[cfg(target_pointer_width = "32")]
-    let length = buf.len() as u32;
+    let length = buf.len() as usize;
 
     let (buffer_id, buffer_ptr) = BUFFERS.with(|rc| {
         let buffers = &mut *rc.borrow_mut();
@@ -163,7 +163,7 @@ pub unsafe fn new_array_buffer_copy(
     #[cfg(target_pointer_width = "64")]
     let length = buf.len();
     #[cfg(target_pointer_width = "32")]
-    let length = buf.len() as u32;
+    let length = buf.len() as usize;
 
     let raw = q::JS_NewArrayBufferCopy(ctx, buf.as_ptr(), length);
     let obj_ref = QuickJsValueAdapter::new(
@@ -216,7 +216,7 @@ pub unsafe fn detach_array_buffer_buffer(
         #[cfg(target_pointer_width = "64")]
         let mut len: usize = 0;
         #[cfg(target_pointer_width = "32")]
-        let mut len: u32 = 0;
+        let mut len: usize = 0;
 
         let ptr = q::JS_GetArrayBuffer(ctx, &mut len, *array_buffer.borrow_value());
 
@@ -268,7 +268,7 @@ pub unsafe fn get_array_buffer_buffer_copy(
         #[cfg(target_pointer_width = "64")]
         let mut len: usize = 0;
         #[cfg(target_pointer_width = "32")]
-        let mut len: u32 = 0;
+        let mut len: usize = 0;
 
         let ptr = q::JS_GetArrayBuffer(ctx, &mut len, *array_buffer.borrow_value());
